task4-7/requirements.txt
fastapi
uvicorn
playwright
python-dotenv
psycopg2-binary